Test QLDG

+++ DAO/QLDG_DAO.cs
@@ -20,30 +20,37 @@ namespace QuanLyThuVien.DAO
         {
             if (GetData("select* from ACCOUNT where MaDocGia = '" + acc.MaDocGia + "'").Rows.Count > 0)
                 return false;
+            else if (GetData("select* from ACCOUNT where TenDangNhap = '" + acc.TenDangNhap + "'").Rows.Count > 0)
+                return false;
             string sql = "INSERT INTO ACCOUNT(MaDocGia, TenDangNhap, MatKhau, Quyen)VALUES (";
-            sql += "N'" + acc.MaDocGia + "',N'" + acc.TenDangNhap + "','" + acc.MatKhau + "','" + 0 + "')";
+            sql += "N'" + acc.MaDocGia + "',N'" + acc.TenDangNhap + "','" + 123 + "','" + 0 + "')";
             Excute(sql);
             string Sql = "INSERT INTO DOCGIA (MaDocGia, TenDangNhap, HoTen, DiaChi,GioiTinh,NamSinh) VALUES (";
             Sql += "N'" + acc.MaDocGia + "',N'" + acc.TenDangNhap + "',N'" + dg.HoTen + "',N'" + dg.DiaChi + "',N'" + dg.GioiTinh + "','" + dg.NamSinh + "')";
             Excute(Sql);
             return true;
         }
-        public void XoaDG(string mS)
+        public bool XoaDG(string mS)
         {
+            if (GetData("select* from ACCOUNT where MaDocGia = '" + mS + "'").Rows.Count <= 0)
+                return false;
             Excute("delete from PHIEUTRA where MaDocGia = '" + mS + "'");
             Excute("delete from PHIEUMUON where MaDocGia = '" + mS + "'");
             Excute("delete from ACCOUNT where MaDocGia = '" + mS + "'");
             Excute("delete from DOCGIA where MaDocGia = '" + mS + "'");
+            return true;
         }
 
         public void SuaDG(DocGia_DTO dg, Account_DTO acc)
         {
+            
             string Sql = string.Format("update ACCOUNT set  TenDangNhap = N'{0}' where MaDocGia = '{1}'",
                 acc.TenDangNhap, acc.MaDocGia);
             Excute(Sql);
             string sql = string.Format("update DOCGIA set  TenDangNhap = N'{0}', HoTen = N'{1}', GioiTinh = N'{2}', NamSinh = '{3}', DiaChi = N'{4}'  where MaDocGia = '{5}'",
                 acc.TenDangNhap, dg.HoTen, dg.GioiTinh, dg.NamSinh, dg.DiaChi, acc.MaDocGia);
             Excute(sql);
+            
         }
         public DataTable TimDG(string _timkiem, string _loaitk)
         {
